Changed segmentation collinear check precision from float to double.

+++ sample_consensus/include/pcl/sample_consensus/impl/sac_model_cone.hpp
@@ -87,7 +87,7 @@ pcl::SampleConsensusModelCone<PointT, PointNT>::computeModelCoefficients (
   Eigen::Vector4f ortho31 = n3.cross3(n1);
 
   float denominator = n1.dot(ortho23);
-  if (std::abs(denominator) < Eigen::NumTraits<float>::dummy_precision ())
+  if (std::abs(denominator) < Eigen::NumTraits<double>::dummy_precision ())
   {
     PCL_ERROR ("[pcl::SampleConsensusModelCone::computeModelCoefficients] Impossible to compute stable model with these points.\n");
     return (false);
